+++ testedApplication/src/main/java/com/zalas/masterthesis/application/redundantcomponents/FindOneRedundantComponent.java
@@ -3,5 +3,5 @@ package com.zalas.masterthesis.application.redundantcomponents;
 import com.zalas.masterthesis.application.model.ProductCategory;
 
 public interface FindOneRedundantComponent {
-    ProductCategory findOne(long id);
+    ProductCategory findOne(int id);
 }
